requirements.txt
python-dotenv
pydantic
requests
GitPython
pytest
openai
langchain-openai
langchain_community
streamlit
langgraph